commit on 2022年 5月14日 週六 12時51分01秒 CST

+++ delete.bash
@@ -1,33 +1,35 @@
 function delete () { 
-
+    cd ${IDKDIR}/gitprocess/ 2>/dev/null
+    if [[ $? -eq 1 ]];then
+        echo "nothing to delete"
+        exit 0
+    fi
     declare -a dlist
-    dlist=$(cd ~/Desktop/.gitprocess/ ; find . -name 'gitprocessof*'| sed -e 's/^..//' -e 's/gitprocessof//' -e 's/.bash//p')
+    dlist=$(find . -name 'gitprocessof*'| sed -e 's/^..//' -e 's/gitprocessof//' -e 's/.bash//p')
     dlist+=" all"
-    dlist+=" back"
+    
 
     PS3="proj name: "
     select p in ${dlist}
     do
         if [[ $p == 'all' ]];then
 
-            for i in $(cd ~/Desktop/.gitprocess/ ; find . -name 'gitprocessof*'| sed -e 's/^..//' -e 's/gitprocessof//' -e 's/.bash//p')
+            for i in $(find . -name 'gitprocessof*'| sed -e 's/^..//' -e 's/gitprocessof//' -e 's/.bash//p')
                 do
-                    cd ~/Desktop/.gitprocess/
                     echo "start deleting..."
                     rm -rf dir gitprocessof$i 
                 done
 
             break
         
-        elif [[ $p == "back" ]];then
+        elif [[ $REPLY == "b/" ]];then
             exit 88
 
         else
             
-            cd ~/Desktop/.gitprocess/
             echo "start deleting..." 
             rm -rf dir gitprocessof$p 
-            cd ~/Desktop/.gitprocess/gitprocessof$p 2>/dev/null
+            cd ${IDKDIR}/gitprocess/gitprocessof$p 2>/dev/null
 
             if [[ $? -eq 1 || $REPLY == '' ]];then   
                 echo "delete fail,project name non-found,choose again"
